ch 7 nest: intro into paras, parallel efficiency, zone2 fig fix
The introductory paragrpah is huge, so it is now broken into
several pieces.

The portion about parallel efficiency is now less concerned with
"ours is better than other models" and more accurately stated as
what happens.

The zone2 figure had two problems, this fixes the second one
```
\ref{figure:zone12}
```
The `figure:`part was still missing.